refactor services and api
