+++ api/src/api/init.rs
@@ -1,10 +1,5 @@
-mod job;
-mod status;
-mod verified_programs;
-mod verify_async;
-mod verify_sync;
 use crate::db::DbClient;
-use crate::routes::{
+use super::routes::{
     job::get_job_status, status::verify_status, verified_programs::get_verified_programs_list,
     verify_async::verify_async, verify_sync::verify_sync,
 };
@@ -12,10 +7,8 @@ use axum::{
     error_handling::HandleErrorLayer,
     http::{Method, StatusCode},
     routing::{get, post},
-    BoxError, Json, Router,
+    BoxError, Router,
 };
-use serde_json::{json, Value};
-use std::sync::OnceLock;
 use std::time::Duration;
 use tower::{buffer::BufferLayer, limit::RateLimitLayer, ServiceBuilder};
 use tower_governor::{
@@ -28,7 +21,10 @@ use tower_http::{
 };
 use tracing::Level;
 
-pub fn create_router(db: DbClient) -> Router {
+use super::index::index;
+
+
+pub fn initialize_router(db: DbClient) -> Router {
     let error_handler = || {
         ServiceBuilder::new().layer(HandleErrorLayer::new(|err: BoxError| async move {
             (
@@ -105,44 +101,4 @@ pub fn create_router(db: DbClient) -> Router {
         )
         .layer(trace_layer)
         .with_state(db)
-}
-
-static INDEX_JSON: OnceLock<Value> = OnceLock::new();
-
-fn index() -> Json<Value> {
-    let value = INDEX_JSON.get_or_init(||
-        json!({
-            "endpoints": [
-                {
-                    "path": "/verify",
-                    "method": "POST",
-                    "description": "Verify a program",
-                    "params" : {
-                        "repo": "Git repository URL",
-                        "program_id": "Program ID of the program in mainnet",
-                        "commit": "(Optional) Commit hash of the repository. If not specified, the latest commit will be used.",
-                        "lib_name": "(Optional) If the repository contains multiple programs, specify the name of the library name of the program to build and verify.",
-                        "bpf_flag": "(Optional)  If the program requires cargo build-bpf (instead of cargo build-sbf), as for an Anchor program, set this flag.",
-                        "base_image": "(Optional) Base docker image to use for building the program.",
-                        "mount_path": "(Optional) Mount path for the repository.",
-                        "cargo_args": "(Optional) Cargo args to pass to the build command. It should be Vector of strings."
-                    },
-                },
-                {
-                    "path": "/status/:address",
-                    "method": "GET",
-                    "description": "Check the verification status of a program by its address",
-                    "params": {
-                        "address": "Address of the mainnet program to check the verification status"
-                    }
-                },
-                {
-                    "path": "/verified-programs",
-                    "method": "GET",
-                    "description": "Get the list of verified programs"
-                }
-            ]
-        })
-    );
-    Json(value.clone())
 }
