+++ api/src/api/mod.rs
@@ -1,0 +1,5 @@
+pub mod index;
+pub mod init;
+pub mod routes;
+
+pub use init::initialize_router;
