+++ api/src/api/routes/mod.rs
@@ -1,0 +1,5 @@
+pub mod job;
+pub mod status;
+pub mod verified_programs;
+pub mod verify_async;
+pub mod verify_sync;
