+++ api/src/services/mod.rs
@@ -1,0 +1,6 @@
+pub mod misc;
+pub mod onchain;
+pub mod verification;
+
+pub use misc::get_repo_url;
+pub use onchain::get_on_chain_hash;
